Moved the fetching from the API to another file and run it in index.js and made the /matches route especially for GET data in my frontend
